+++ gFunctionDatabase/Management/application.py
@@ -1,9 +1,11 @@
 # Jack C. Cook
 # Thursday, September 16, 2021
+import warnings
 
 from scipy.interpolate import interp1d, lagrange
 import matplotlib.pyplot as plt
 import numpy as np
+import pygfunction as gt
 
 
 class GFunction:
@@ -16,14 +18,15 @@ class GFunction:
         self.g_lts: dict = g_lts  # g-functions (LTS) keyed by height
         # ln(t/ts) values that apply to all the heights
         self.log_time: list = log_time
-        self.bore_locations: list = bore_locations  # (x, y) coordinates of boreholes
+        # (x, y) coordinates of boreholes
+        self.bore_locations: list = bore_locations
         # self.time: dict = {}  # the time values in years
 
         # an interpolation table for B/H ratios, D, r_b (used in the method
         # g_function_interpolation)
         self.interpolation_table: dict = {}
 
-    def g_function_interpolation(self, B_over_H: float, kind='cubic'):
+    def g_function_interpolation(self, B_over_H: float, kind='default'):
         """
         Interpolate a range of g-functions for a specific B/H ratio
         Parameters
@@ -49,6 +52,43 @@ class GFunction:
         # the g-functions are stored in a dictionary based on heights, so an
         # equivalent height can be found
         H_eq = 1 / B_over_H * self.B
+
+        # Determine if we are out of range and need to extrapolate
+        height_values = list(self.g_lts.keys())
+        if min(height_values) <= H_eq <= max(height_values):
+            fill_value = ''
+        else:
+            fill_value = 'extrapolate'
+            warnings.warn('Extrapolation is being used.')
+
+        # if the interpolation kind is default, use what we know about the
+        # accuracy of interpolation to choose a technique
+        if kind == 'default':
+            num_curves = len(height_values)
+            if num_curves >= 5:
+                kind = 'cubic'
+            elif num_curves >= 3:
+                kind = 'quadratic'
+            elif num_curves == 2:
+                kind = 'linear'
+            else:
+                raise ValueError(
+                    'Interpolation requires two g-function curves.')
+
+        # Automatically adjust interpolation if necessary
+        # Lagrange also needs 2
+        interpolation_kinds = {'linear': 2, 'quadratic': 3, 'cubic': 4,
+                               'lagrange': 2}
+        curves_by_kind = {2: 'linear', 3: 'quadratic', 4: 'cubic'}
+        if len(height_values) < 2:
+            raise ValueError('Interpolation requires two g-function curves.')
+        else:
+            # If the number of required curves for the interpolation type is
+            # greater than what is available, reduce the interpolation type
+            required_curves = interpolation_kinds[kind]
+            if required_curves > len(height_values):
+                kind = curves_by_kind[len(height_values)]
+
         # if the interpolation table is not yet know, build it
         if len(self.interpolation_table) == 0:
             # create an interpolation for the g-function which takes the height
@@ -66,7 +106,7 @@ class GFunction:
                 if kind == 'lagrange':
                     f = lagrange(x, y)
                 else:
-                    f = interp1d(x, y, kind=kind)
+                    f = interp1d(x, y, kind=kind, fill_value=fill_value)
                 self.interpolation_table['g'].append(f)
             # create interpolation tables for 'D' and 'r_b' by height
             keys = list(self.r_b_values.keys())
@@ -84,13 +124,13 @@ class GFunction:
                 rb_f = lagrange(height_values, rb_values)
             else:
                 # interpolation function for rb values by H equivalent
-                rb_f = interp1d(height_values, rb_values, kind=kind)
+                rb_f = interp1d(height_values, rb_values, kind=kind, fill_value=fill_value)
             self.interpolation_table['rb'] = rb_f
             try:
                 if kind == 'lagrange':
                     D_f = lagrange(height_values, D_values)
                 else:
-                    D_f = interp1d(height_values, D_values, kind=kind)
+                    D_f = interp1d(height_values, D_values, kind=kind, fill_value=fill_value)
                 self.interpolation_table['D'] = D_f
             except:
                 pass
@@ -133,6 +173,25 @@ class GFunction:
             g_function_corrected.append(g_function[i] - np.log(rb_star / rb))
         return g_function_corrected
 
+    def correct_coordinates(self, B: float):
+        # scale the field by the ratio
+        scale = B / self.B
+
+        new_locations = []
+        for i in range(len(self.bore_locations)):
+            _x = self.bore_locations[i][0]
+            _y = self.bore_locations[i][1]
+            if _x == 0.0 and _y == 0.0:
+                new_locations.append([_x, _y])
+            elif _x == 0.0:
+                new_locations.append([_x, _y*scale])
+            else:
+                x = _x * scale
+                y = _y * scale
+                new_locations.append([x, y])
+
+        return new_locations
+
     def visualize_g_functions(self):
         """
         Visualize the g-functions.
@@ -152,7 +211,7 @@ class GFunction:
 
         for key in keys:
             ax.plot(self.log_time, self.g_lts[key],
-                    label=str(int(self.B)) + '/' + str(key))
+                    label=str(int(self.B)) + '/' + str(int(key)))
             x_n = self.log_time[-1]
             y_n = self.g_lts[key][-1]
             if key == 8:
@@ -164,9 +223,11 @@ class GFunction:
 
         handles, labels = ax.get_legend_handles_labels()
 
+        # legend = fig.legend(handles=handles, labels=labels,
+        #                     title='B/H'.rjust(5) + '\nLibrary',
+        #                     bbox_to_anchor=(1, 1.0))
         legend = fig.legend(handles=handles, labels=labels,
-                            title='B/H'.rjust(5) + '\nLibrary',
-                            bbox_to_anchor=(1, 1.0))
+                            title='B/H'.rjust(5), bbox_to_anchor=(1, 1.0))
         fig.gca().add_artist(legend)
 
         ax.set_ylabel('g')
@@ -185,10 +246,10 @@ class GFunction:
         **fig, ax**
             Figure and axes information.
         """
-        fig, ax = plt.subplots(figsize=(3.5, 5))
-
         if bore_locations is None:
             bore_locations = self.bore_locations
+
+        fig, ax = plt.subplots(figsize=(3.5, 5))
 
         x, y = list(zip(*bore_locations))
 
@@ -201,3 +262,143 @@ class GFunction:
         fig.tight_layout()
 
         return fig, ax
+
+    @staticmethod
+    def visualize_area_and_constraints(perimeter: list, coordinates: list,
+                                       no_go=None):
+        """
+        Visualize the (x,y) coordinates, and no go zone.
+
+        Returns
+        -------
+        **fig, ax**
+            Figure and axes information.
+        """
+        fig, ax = plt.subplots()
+
+        perimeter = perimeter + [perimeter[0]]
+
+        x, y = list(zip(*coordinates))
+
+        ax.scatter(x, y)
+        _x, _y = list(zip(*perimeter))
+        ax.plot(_x, _y, 'g')
+        if no_go is not None:
+            no_go = no_go + [no_go[0]]
+            _x, _y = list(zip(*no_go))
+            ax.plot(_x, _y, 'r')
+
+        ax.set_xlabel('x (m)')
+        ax.set_ylabel('y (m)')
+
+        fig.tight_layout()
+
+        return fig, ax
+
+    @staticmethod
+    def configure_database_file_for_usage(data) -> dict:
+        """
+        This method is called upon initialization of the object.
+        Read the cpgfunction output dictionary into the borefield class for easy
+        access of information
+        Parameters
+        ----------
+        data: dict
+            A dictionary which is in the output format of cpgfunction.
+        Returns
+        -------
+            a dictionary for input to this object
+        """
+        log_time = data['logtime']
+
+        bore_locations = data[
+            'bore_locations']  # store the bore locations in the object
+        g_values: dict = data['g']  # pull the g-functions into the g_values
+        g_tmp: dict = {}  # a temporary g-function dictionary that might be out of order
+        Ds_tmp: dict = {}  # a temporary burial depth dictionary that may be out of order
+        r_bs_tmp: dict = {}  # the borehole radius dictionary that may be out of order
+        t_tmp: dict = {}
+        for key in g_values:
+            # do the g-function dictionary
+            key_split = key.split('_')
+            # get the current height
+            height = float((key_split[1]))
+            # create a g-function list associated with this height key
+            g_tmp[height] = g_values[key]
+            # create a r_b value associated with this height key
+            r_b = float(key_split[2])
+            r_bs_tmp[height] = r_b
+            try:  # the D value is recently added to the key value for the saved g-functions computed
+                D = float(key_split[3])
+                Ds_tmp[height] = D
+            except:
+                pass
+
+            # do the time dictionary
+            time_arr: list = []
+            for _, lntts in enumerate(log_time):
+                alpha = 1.0e-06
+                t_seconds = height ** 2 / 9 / alpha * np.exp(lntts)
+                t_year = t_seconds / 60 / 24 / 365
+                time_arr.append(t_year)
+            t_tmp[height] = time_arr
+
+        B = float(list(g_values.keys())[0].split('_')[
+                           0])  # every B-spacing should be the same for each file
+
+        keys = sorted(list(g_tmp.keys()), key=int)  # sort the heights in order
+
+        g = {key: g_tmp[key] for key in
+                  keys}  # fill the g-function dictionary with sorted heights
+        try:
+            Ds = {key: Ds_tmp[key] for key in
+                       keys}  # fill the burial depth dictionary with sorted heights
+        except:
+            # if there's no D provided, make it 2
+            Ds = {key: 2. for key in keys}
+        r_bs = {key: r_bs_tmp[key] for key in keys}
+        time = {key: t_tmp[key] for key in
+                     keys}  # fill the time array for yearly points
+
+        geothermal_g_input = {}
+        geothermal_g_input['B'] = B
+        geothermal_g_input['r_b_values'] = r_bs
+        geothermal_g_input['D_values'] = Ds
+        geothermal_g_input['g_lts'] = g
+        geothermal_g_input['log_time'] = log_time
+        geothermal_g_input['bore_locations'] = bore_locations
+
+        return geothermal_g_input
+
+
+def uniform_temperature_uniform_segment_lengths(D_values: list, H_values: list,
+                                                r_b_values: list, B: float,
+                                                log_time: list, alpha: float,
+                                                coordinates: list,
+                                                nSegments: list):
+    # Compute a uniform borehole wall temperature g-function with uniform
+    # segment lengths
+
+    g_data: dict = {'bore_locations': coordinates, 'g': {}, 'logtime': log_time}
+
+    for i in range(len(D_values)):
+        D = D_values[i]
+        H = H_values[i]
+        r_b = r_b_values[i]
+        nq = nSegments[i]
+        # g-Function calculation options
+        options = {'nSegments': nq, 'disp': False}
+        t_s = H ** 2 / (9 * alpha)
+        t = np.exp(np.array(log_time)) * t_s
+        boreField = []
+        for j in range(len(coordinates)):
+            x, y = coordinates[j]
+            borehole = gt.boreholes.Borehole(H, D, r_b, x, y)
+            boreField.append(borehole)
+        gfunc_uniform_T = gt.gfunction.gFunction(
+            boreField, alpha, time=t, boundary_condition='UBWT',
+            options=options)
+        key_value = '{}_{}_{}_{}'.format(B, H, r_b, D)
+        g_data['g'][key_value] = gfunc_uniform_T.gFunc.tolist()
+
+    return g_data